requirements.txt
pandas
numpy
psycopg2-binary
sqlalchemy
requests
beautifulsoup4
feedparser
boto3
s3fs
pyarrow
mlflow
scikit-learn
pyspark==3.5.3
delta-spark==3.2.0
chromadb
sentence-transformers
deltalake
python-dotenv
tqdm